* conv_princ.c (krb5_425_conv_principal): Cast argument to tolower 	to int.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13575 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/conv_princ.c
@@ -319,7 +319,7 @@ krb5_425_conv_principal(context, name, instance, realm, princ)
 		  if (domain) {
 		      for (cp = domain; *cp; cp++)
 			  if (isupper((int) (*cp)))
-			      *cp = tolower(*cp);
+			      *cp = tolower((int) *cp);
 		      strncat(buf, ".", sizeof(buf) - 1 - strlen(buf));
 		      strncat(buf, domain, sizeof(buf) - 1 - strlen(buf));
 		      krb5_xfree(domain);
